Arduino/requirements.txt
mido==1.2.8
pygame==1.9.3
pyserial==3.4
